Separate RPC from hooks
