+++ waspc/data/Generator/templates/sdk/wasp/client/operations/rpc.ts
@@ -1,0 +1,38 @@
+import { type Route } from "wasp/client";
+import type {
+  _Awaited,
+  _ReturnType 
+} from "wasp/universal/types"
+
+// PRIVATE API (but should maybe be public, users define values of this type)
+export type Query<Input, Output> = QueryFunction<Input, Output> & QueryMetadata
+
+// PRIVATE API (but should maybe be public, users define values of this type)
+export type Action<Input, Output> = ClientOperation<Input, Output>
+
+// PRIVATE API
+export type QueryFunction<Input, Output> = ClientOperation<Input, Output>
+
+// PRIVATE API
+export type QueryMetadata = {
+  queryCacheKey: string[]
+  route: Route
+}
+
+// PRIVATE API (needed in SDK)
+export type OperationRpcFor<BackendOperation extends GenericBackendOperation> =
+  Parameters<BackendOperation> extends []
+    ? ClientOperation<void, _Awaited<_ReturnType<BackendOperation>>>
+    : ClientOperation<
+        Parameters<BackendOperation>[0],
+        _Awaited<_ReturnType<BackendOperation>>
+      >
+
+// PRIVATE API (but should maybe be public, users use values of this type)
+// Read this to understand the type: https://github.com/wasp-lang/wasp/pull/1090#discussion_r1159732471
+export type ClientOperation<Input, Output> = [Input] extends [never]
+  ? (args?: unknown) => Promise<Output>
+  : (args: Input) => Promise<Output>;
+
+// PRIVATE API (needed in SDK)
+export type GenericBackendOperation = (args: never, context: any) => unknown
